+++ resources/js/Stores/WelcomeStore.js
@@ -1,11 +1,12 @@
 import { defineStore } from "pinia";
+import {ref} from "vue";
 
 export let useWelcomeStore = defineStore('welcomeStore', {
     state() {
         return {
             showLogin: Boolean,
             showRegister: Boolean,
-            showOverlay: Boolean,
+            showOverlay: true,
         };
     },